Upload My Code
I am currently using Visual Studio 2019 to code the program. But I'm not sure whether I should upload the folder ".vs" and "Debug". So that's it for now. I wonder how to "connect" Visual Studio with GitHub.
PS: Chinese is my mother tongue, thus the notes were written in Chinese.